Use permanent storage for database.db
Follow-up PR after [#2](https://huggingface.co/spaces/ttseval/TTS-Arena/discussions/2) cc @mrfakename @hysts 

With this PR, no need to re-download the database if the persistent storage is enabled. This should avoid loosing data because restarts. In case we disable permanent storage in the future (for any reason), the app will still work and default to the local storage as done at the moment.